% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{print.smgeecure}
\alias{print.smgeecure}
\title{Print smgeecure object}
\usage{
\method{print}{smgeecure}(fit)
}
\arguments{
\item{fit}{an object of smgeecure}
}
\description{
Output of \code{smgeecure} object.
}
