% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.eRes}
\alias{SMC.AFT.eRes}
\title{Survival function for latency}
\usage{
SMC.AFT.eRes(tobs, delta, X, beta, gm = rep(1, length(yobs)), intercept = TRUE)
}
\arguments{
\item{tobs}{right censored data which is the follow up time.}

\item{delta}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring.}

\item{X}{a vector or matrix of covariates corresponding to latency part.}

\item{gm}{the conditional expectation of the cure status given the observed data and the current.
estimates of the parameters in each E-step. The default \code{gm=rep(1,length(yobs))}.}

\item{bet}{unknown parameters corresponding to latency part.}
}
\description{
Estimate the survival function of latency part used in the M-step of the EM algorithm for the AFT mixture cure model.
(this convex objective function that will be minimized in each EM iteration)
}
