% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.fit}
\alias{SMC.AFT.fit}
\title{Semiparametric accelerated failure time mixture cure model}
\usage{
SMC.AFT.fit(yobs, delta, X, Z, incidence = c("logit"), em.maxit = 50)
}
\arguments{
\item{yobs}{right censored data which is the follow up time.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a vector or matrix of covariates corresponding to latency part.}

\item{Z}{a vector or matrix of covariates corresponding to incidence part.}

\item{incidence}{specifies the link in incidence part. The "logit", "probit" or
complementary loglog ("cloglog") links are available. By default incidence = c("logit").}

\item{em.maxit}{specifies the maximum iteration number. If the convergence criterion
is not met, the EM iteration will be stopped after emmax iterations and the estimates
will be based on the last maximum likelihood iteration. The default \code{em.maxit = 100}.}
}
\description{
Fit semiparametric accelerated failure time mixture cure model by the EM algorithm.
The details about the method can be found in Zhang and Peng (2007).
This function is indeed the same as that in Cai et al. (2012), but with different notations and sub-functions
}
