% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.GEE.control}
\alias{SMC.AFT.GEE.control}
\title{Auxiliary for Controlling AFT GEE Fitting}
\usage{
SMC.AFT.GEE.control(
  emmax = 15,
  emeps = 1e-04,
  emtrace = FALSE,
  emmax.smcure = 50,
  maxiter.latency = 15,
  eps.latency = 1e-04,
  trace.latency = FALSE
)
}
\arguments{
\item{emmax}{max number of iteration.}

\item{emeps}{relative error tolerance.}

\item{emtrace}{a binary variable, determine whether to display output for each iteration.}

\item{emmax.smcure}{max number of iteration.}

\item{maxiter.latency}{max number of iteration.}

\item{eps.latency}{relative error tolerance.}

\item{trace.latency}{a binary variable, determine whether to display output for each iteration.}
}
\description{
Auxiliary function as user interface for \code{SMC.AFT.GEE} fitting.
Only used when calling 'SMC.AFT.GEE' or 'SMC.AFT.GEE.fit'.
}
\seealso{
\code{\link{SMC.AFT.GEE}}
}
