% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.GEE.fit}
\alias{SMC.AFT.GEE.fit}
\title{Sub-fitting functions for semiparametric accelerated failure time mixture cure model}
\usage{
SMC.AFT.GEE.fit(
  tobs,
  delta,
  id,
  X,
  Z,
  corstr = c("ind", "ar1", "ex"),
  Intercept = FALSE,
  control = SMC.AFT.GEE.control()
)
}
\arguments{
\item{tobs}{right censored data which is the follow up time.}

\item{delta}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{X}{a vector or matrix of covariates corresponding to latency part.}

\item{Z}{a vector or matrix of covariates corresponding to incidence part.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted:
"independence", "exchangeable", "ar1".}

\item{control}{a list of iteration and algorithmic constants. See \code{SMC.AFT.GEE.control} for their names and default values.}
}
\description{
Sub-functioins for fitting semiparametric accelerated failure time mixture cure model by the ES algorithm.
}
