% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.GEE}
\alias{SMC.AFT.GEE}
\title{Semiparametric marignal accelerated failure time mixture cure model}
\usage{
SMC.AFT.GEE(
  tobs,
  delta,
  id,
  X,
  Z,
  corstr = c("ind", "ar1", "ex"),
  Var = TRUE,
  nboot = 100,
  nwrong.rate = 0.4,
  Intercept = FALSE,
  control = SMC.AFT.GEE.control()
)
}
\arguments{
\item{tobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{X}{a vector or matrix of covariates corresponding to latency part}

\item{Z}{a vector or matrix of covariates corresponding to incidence part}

\item{corstr}{a character string specifying the correlation structure. The following are permitted:
"independence"(\code{ind}), "exchangeable"(\code{ex}), "ar1"(\code{ar1}).}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{nwrong.rate}{the acceptable rate of wrong fitted bootstap samples.}

\item{Intercept}{Whether to include an intercept term in the latency part}

\item{control}{a list of iteration and algorithmic constants. ee \code{SMC.AFT.GEE.control} for their names and default values.
These can also be set as arguments to geese itself.}
}
\description{
Fit semiparametric marignal accelerated failure time mixture cure model by the EM algorithm.
The details about the method can be found in Zhang and Peng (2007).
This function is indeed the same as that in Cai et al. (2012), but with different notations and sub-functions
}
