% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{SMC.AFT.rank}
\alias{SMC.AFT.rank}
\title{Rank function}
\usage{
SMC.AFT.rank(bet, yobs, delta, X, w, intercept = TRUE)
}
\arguments{
\item{bet}{unknown parameters corresponding to latency part.}

\item{yobs}{right censored data which is the follow up time.}

\item{delta}{the censoring indicator, normally 0 = event of interest happens, and 0 = censoring.}

\item{X}{a vector or matrix of covariates corresponding to latency part.}

\item{w}{conditional probability of the individual remaining uncured.}
}
\description{
Rank estimating equation used in the M-step of the EM algorithm for the AFT mixture cure model.
(this convex objective function that will be minimized in each EM iteration)
}
