% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{smgeecure}
\alias{smgeecure}
\title{smgeecure: Mixture Cure Models with Auxiliary Subgroup Survival Probabilities.}
\usage{
smgeecure(
  formula,
  cureform,
  id,
  data,
  model = c("aft", "ph"),
  corstr = c("independence", "exchangeable", "ar1"),
  Var = TRUE,
  nboot = 100,
  stdz = TRUE,
  esmax = 20,
  eps = 1e-04
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring. It is used to specify the covariate effects on the failure time of uncured subjects. See the documentation for \code{survreg} and \code{Surv} in package \code{survival} for details.
The expression to the right of the "~" specifies the effect of covariates on the failure time of uncured patients.}

\item{cureform}{indicator function a formula expression, of the form \code{cureform ~ predictors}. It is used to specify the effects of covariates on the cure rate. A covariate may be used in both \code{formula} and \code{cureform}.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula} and the \code{cureform}.}

\item{model}{specifies the model using in latency part. It can be \code{ph} which represents
the proportional hazards model, or \code{aft} which represents the accelerated failure time model.}

\item{corstr}{a character string specifying the correlation structure. Currently, the following three working correlation structures are available:
\code{independence} for the independent working correlation structure,
\code{exchangeable} for the exchangeable working correlation structure and
\code{ar1} for the first-order auto-regressive structure.}

\item{Var}{a logical value. If it is TRUE, the program returns standard error estimates for beta and gamma.
Otherwise, no standard errors are computed and only coefficient estimates are returned.
By default, \code{Var = TRUE}. Note that for both \code{model="ph"} and \code{"aft"}, the returned
standard error estimators are based on the bootstrap method.}

\item{nboot}{specifies the number of bootstrap samplings. The default value is 100.}

\item{stdz}{a logical value. If it is \code{TRUE}, all the covariates are standardized. Otherwise, non-transformed covariates are used.}

\item{esmax}{specifies the maximum iteration number. If the convergence criterion is not met,
the ES iteration will be stopped after \code{esmax} iterations and the estimates will be based on the
last ES iteration. The default \code{esmax = 100}.}

\item{eps}{specifies the relative tolerance, and is set to \code{1e-06} as the default.
The iterations are considered to be converged when the maximum relative change in the
parameters and likelihood estimates between iterations is less than the value specified.}
}
\value{
An object of class \code{smgeecure} is returned. It can be examined by \code{print.smgeecure()}.
}
\description{
This package features the marginal semi-parametric proportional hazards/accelerated failure time mixture cure models for analyzing clustered survival data with a possible cure fraction.
The underlying methods are based on the paper titled "Marginal Accelerated Failure Time Cure Model for clustered survival data with long-term survivors",
which has been submitted to the journal: Statistical Methods in Medical Research.

Fit the semiparametric marginal proportional hazards mixture cure (PHMC) model or
the semiparametric accelerated failure time mixture cure (AFTMC) model with the generalized
estimating equations (GEE). GEE approach is generalized through the expectation-solution (ES) algorithm
to account for the correlation among the cure statuses and the dependence among the failure
times of uncured patients to improve the estimation efficiency.
}
\details{
Semiparametric marginal PHMC and AFTMC models are considered in this function.
For cure rate, a logistic regression model is employed and the probability of being cured is given
by \code{(1+exp(gamZ))^{(-1)}}.
A covariate can be used either in formula or in cureform or in both. The model parameters are
estimated by the expectation-solution (ES) algorithm and the standard error estimates are obtained
using bootstrap.
}
\examples{
\donttest{
### Be patient, the following examples may take several minutes on a faster computer.

### library
library(smgeecure)
library(geecure)

#=======================================================================#
### Example 1. Fit the marginal semiparametric PHMC model for
###             the tonsil data.

# ---- prepare the data
data(tonsil)
tonsil <- tonsil[-c(141,136,159),]
tonsil$Sex <- ifelse(tonsil$Sex == 1, 0, 1) # 1="Female"
tonsil$Cond <- ifelse(tonsil$Cond == 1, 0, 1) # 0=no disability
tonsil$T <- ifelse(tonsil$T < 4, 0, 1)
tonsil$Grade2 <- ifelse(tonsil$Grade==2,1,0)
tonsil$Grade3 <- ifelse(tonsil$Grade==3,1,0)
table(tonsil$Inst)

# ---- plot the KM survival curve (overall)
pdf("Figure_Tonsil_KM.pdf",width=7,height=6)
plot(survival::survfit(survival::Surv(Time, Status) ~ 1, data = tonsil),
     conf.int = T, mark.time = TRUE, lwd=1.5,
     ylab = "Survival Probability", xlab = "Survival Time (in Days)",
     xlim = c(0,2000), ylim = c(0,1),
     cex.lab = 1,
     xaxt = "n", yaxt = "n"
)
axis(1,seq(0,2000,500),seq(0,2000,500),cex.axis = 1)
axis(2,seq(0,1,0.2),seq(0,1,0.2),cex.axis = 1)
dev.off()

# ---- plot the KM survival curve (stratified)
pdf("Figure_Tonsil_KM_SexTumorsize.pdf",width=7,height=6)
plot(survival::survfit(survival::Surv(Time, Status) ~ Sex + T, data = tonsil),
     conf.int = F,  mark.time = TRUE, lwd=1.5,
     ylab = "Survival Probability", xlab = "Survival Time (in Days)",
     xlim = c(0,2000), ylim = c(0,1),
     lty = c(1,2,3,4), col = c(1,2,3,4),
     xaxt = "n", yaxt = "n", cex.lab=1)
axis(1,seq(0,2000,400),seq(0,2000,400),cex.axis=1)
axis(2,seq(0,1,0.2),seq(0,1,0.2),cex.axis=1)
legend("topright", c("Massive/Female","Massive/Male","Non-massive/Female","Non-massive/Male"),
       cex = 1, col = c(1,2,3,4), lty = c(1,2,3,4))
dev.off()

# ---- plot the log-cumulative survival curve (check ph assumption)
#         Plot BY Sex
# estimates for 1
KM.fit.1 <- survival::survfit(survival::Surv(Time, Status) ~ 1,
                              data = tonsil, subset = (Sex==1))
Infty.1 <- min(KM.fit.1$surv)
surv.1.Latency <- (KM.fit.1$surv-Infty.1) / (1-Infty.1)
time.1.Latency <- KM.fit.1$time
cumu.1.Latency.Nozero <- -log(surv.1.Latency[surv.1.Latency!=0])
time.1.Latency.Nozero <- KM.fit.1$time[surv.1.Latency!=0]
# estimates for 0
KM.fit.0 <- survival::survfit(survival::Surv(Time, Status) ~ 1,
                              data = tonsil, subset = (Sex==0))
Infty.0 <- min(KM.fit.0$surv)
surv.0.Latency <- (KM.fit.0$surv-Infty.0) / (1-Infty.0)
time.0.Latency <- KM.fit.0$time
cumu.0.Latency.Nozero <- -log(surv.0.Latency[surv.0.Latency!=0])
time.0.Latency.Nozero <- KM.fit.0$time[surv.0.Latency!=0]
# plot of log-cumulative function
pdf("Figure_Tonsil_LogCum_Sex.pdf",width=7,height=6)
plot(log(cumu.1.Latency.Nozero)~log(time.1.Latency.Nozero),type="s",lty=1,col="red",
     xlim = c(0,8), ylim=c(-5,1.5), lwd=1.5,
     ylab = "Logarithm of the Cumulative Hazard Function",
     xlab = "Logarithm of the Survival Time",
     xaxt = "n",
     cex.lab=1
)
lines(log(cumu.0.Latency.Nozero)~log(time.0.Latency.Nozero),type="s",lty=2,col="black",lwd=1.5)
axis(1,seq(0,8,2),seq(0,8,2),cex.axis = 1)
legend("bottomright", c("Female","Male"), cex = 1,
       col = c("red","black"), lty = c(1,2))
dev.off()

# ---- plot the log-cumulative survival curve (check ph assumption)
#         Plot BY Sex
#         (Just use uncensored patients)
UncuredPatients <- tonsil$Status==1
pdf("Figure_Tonsil_LogCum_Uncen.pdf",width=7,height=6)
xlim_lower <- 0; xlim_upper <- 1800; xlim_step  <- 300
ylim_lower <- -5; ylim_upper <- 2; ylim_step  <- 1
mylty <- c(1,2); mycol <- c("black","red")
plot(survival::survfit(survival::Surv(Time, Status) ~ Sex,
                       data = tonsil, subset=UncuredPatients,
                       stype=1),
     conf.int = F,  mark.time = FALSE, cumhaz = TRUE, log="xy",
     ylab = "Logarithm of the Cumulative Hazard Function",
     xlab = "Logarithm of the Survival Time (in Days)",
     # xlim = c(xlim_lower,xlim_upper), ylim = c(ylim_lower,ylim_upper),
     lty = mylty, col = mycol, lwd=1.5,
     xaxt = "n", yaxt = "n")
axis(1,c(100,300,600,1500),c(100,300,600,1500))
axis(2,exp(seq(ylim_lower,ylim_upper,ylim_step)),seq(ylim_lower,ylim_upper,ylim_step))
legend("bottomright", c("male","female"), cex = 1.3, #bty="n",
       col = mycol, lty = mylty)
dev.off()

# ---- Fit the data using marginal semi-parametric marginal AFTMC model

# fit marginal semi-parametric marginal AFTMC model (exchangeable correlation)
set.seed(521)
tonsil.aft.gee.ex <- smgeecure(
  formula = Surv(Time, Status) ~ Sex + factor(Grade) + Age + Cond + T,
  cureform = ~ Sex + factor(Grade) + Age + Cond + T, id = tonsil$Inst,
  data = tonsil, model = "aft", corstr = "exchangeable", Var = T, nboot = 100
)
print.smgeecure(tonsil.aft.gee.ex)

# fit marginal semi-parametric AFTMC model (independence correlation)
set.seed(521)
tonsil.aft.gee.ind <- smgeecure(
  formula = Surv(Time, Status) ~ Sex + factor(Grade) + Age + Cond + T,
  cureform = ~ Sex + factor(Grade) + Age + Cond + T, id = tonsil$Inst,
  data = tonsil, model = "aft", corstr = "independence", Var = T, nboot = 100
)
print.smgeecure(tonsil.aft.gee.ind)

# output the results
write.csv(cbind(tonsil.aft.gee.ex$incidence,rbind(NA,tonsil.aft.gee.ex$latency)),
          "tonsil.aft.gee.ex.csv")
write.csv(cbind(tonsil.aft.gee.ind$incidence,rbind(NA,tonsil.aft.gee.ind$latency)),
          "tonsil.aft.gee.ind.csv")

# ---- Fit the data using marginal semi-parametric marginal PHMC model

# fit marginal semi-parametric PHMC model (exchangeable correlation)
set.seed(1)
tonsil.ph.gee.ex <- smgeecure(
  formula = Surv(Time, Status) ~ Sex + factor(Grade) + Age + Cond + T,
  cureform = ~ Sex + factor(Grade) + Age + Cond + T, id = tonsil$Inst,
  data = tonsil, model = "ph", corstr = "exchangeable", Var = T,nboot = 100
)
print.smgeecure(tonsil.ph.gee.ex)

# fit marginal semi-parametric PHMC model (independence correlation)
set.seed(1)
tonsil.ph.gee.ind <- smgeecure(
  formula = Surv(Time, Status) ~ Sex + factor(Grade) + Age + Cond + T,
  cureform = ~ Sex + factor(Grade) + Age + Cond + T, id = tonsil$Inst,
  data = tonsil, model = "ph", corstr = "independence", Var = F, nboot = 100
)
print.smgeecure(tonsil.ph.gee.ind)

# output the results
write.csv(cbind(tonsil.ph.gee.ex$incidence,rbind(NA,tonsil.ph.gee.ex$latency)),
          "tonsil.ph.gee.ex.csv")
write.csv(cbind(tonsil.ph.gee.ind$incidence,rbind(NA,tonsil.ph.gee.ind$latency)),
          "tonsil.ph.gee.ind.csv")


#=======================================================================#
### Example 2. Fit the semiparametric marginal PHMC model for
###             the bone marrow transplantation data.

# ---- prepare the data
data(bmt)
bmt$g <- factor(bmt$g, label = c("ALL", "AML low risk","AML high risk"))
bmt$Z8 <- factor(bmt$Z8, label = c("Otherwise", "FAB"))

# ---- plot the KM survival curve (overall and stratified)
pdf("Figure_bmt_KM.pdf",width=7,height=6)
plot(survival::survfit(survival::Surv(T2, d3) ~ 1, data = bmt),
     conf.int = T, mark.time = TRUE, lwd=1.5,
     ylab = "Survival Probability", xlab = "Survival Time (in Days)",
     xlim = c(0,2800), ylim = c(0,1),
     xaxt = "n", yaxt = "n",
     cex.lab = 1
)
axis(1,seq(0,2800,400),seq(0,2800,400),cex.axis = 1)
axis(2,seq(0,1,0.2),seq(0,1,0.2),cex.axis = 1)
dev.off()

# ---- plot the KM survival curve (overall and stratified)
pdf("Figure_bmt_KM_g.pdf",width=7,height=6)
plot(survival::survfit(survival::Surv(T2, d3) ~ factor(g), data = bmt),
     conf.int = F,  mark.time = TRUE, lwd=1.5,
     ylab = "Survival Probability", xlab = "Survival Time (in Days)",
     xlim = c(0,2800), ylim = c(0,1),
     lty = c(1,2,3), col = c(1,2,3),
     xaxt = "n", yaxt = "n", cex.lab=1)
axis(1,seq(0,2800,400),seq(0,2800,400),cex.axis=1)
axis(2,seq(0,1,0.2),seq(0,1,0.2),cex.axis=1)
legend("topright", c("ALL","AML low risk","AML high risk"),
       cex = 1, col = c(1,2,3), lty = c(1,2,3))
dev.off()


# ---- plot the log-cumulative survival curve (check ph assumption)
#         Plot by g's two dummy factors
pdf("Figure_bmt_LogCum_g.pdf",width=10,height=5)
par(mfrow = c(1,2))
groups.names <- c("AML low risk","AML high risk")
groups <- cbind(bmt$g==groups.names[1],bmt$g==groups.names[2])
for(igroup in 1:ncol(groups)){
  # estimates for 1
  KM.fit.1 <- survival::survfit(survival::Surv(T2, d3) ~ 1,
                                data = bmt, subset = groups[,igroup])
  Infty.1 <- min(KM.fit.1$surv)
  surv.1.Latency <- (KM.fit.1$surv-Infty.1) / (1-Infty.1)
  time.1.Latency <- KM.fit.1$time
  cumu.1.Latency.Nozero <- -log(surv.1.Latency[surv.1.Latency!=0])
  time.1.Latency.Nozero <- KM.fit.1$time[surv.1.Latency!=0]
  # estimates for 0
  KM.fit.0 <- survival::survfit(survival::Surv(T2, d3) ~ 1,
                                data = bmt, subset = !groups[,igroup])
  Infty.0 <- min(KM.fit.0$surv)
  surv.0.Latency <- (KM.fit.0$surv-Infty.0) / (1-Infty.0)
  time.0.Latency <- KM.fit.0$time
  cumu.0.Latency.Nozero <- -log(surv.0.Latency[surv.0.Latency!=0])
  time.0.Latency.Nozero <- KM.fit.0$time[surv.0.Latency!=0]
  # plot of log-cumulative function
  plot(log(cumu.1.Latency.Nozero)~log(time.1.Latency.Nozero),type="s",lty=1,col="red",
       xlim = c(0,8), ylim=c(-4.5,1.5), lwd=1.5,
       ylab = "Logarithm of the Cumulative Hazard Function",
       xlab = "Logarithm of the Survival Time",
       main=groups.names[igroup],
       xaxt = "n",
       cex.lab = 1,cex.main=1
  )
  lines(log(cumu.0.Latency.Nozero)~log(time.0.Latency.Nozero),type="s",lty=2,col="black",lwd=1.5)
  axis(1,seq(0,8,2),seq(0,8,2),cex.axis = 1)
  legend("bottomright", c(groups.names[igroup],"Other"), cex = 1,
         col = c("red","black"), lty = c(1,2))
}
dev.off()


# ---- Fit the data using marginal semi-parametric marginal PHMC model

# fit marginal semi-parametric PHMC model (exchangeable correlation)
set.seed(1)
bmt.ph.gee.ex <- smgeecure(
  formula = Surv(T2, d3) ~ factor(g) + Z8, cureform = ~ factor(g) + Z8,
  id = bmt$Z9, data = bmt, model = "ph", corstr = "exchangeable",
  Var = T,nboot = 100, esmax = 100, eps = 1e-06
)
print.smgeecure(bmt.ph.gee.ex)

# fit marginal semi-parametric PHMC model (independence correlation)
set.seed(1)
bmt.ph.gee.ind <- smgeecure(
  formula = Surv(T2, d3) ~ factor(g) + Z8, cureform = ~ factor(g) + Z8,
  id = bmt$Z9, data = bmt, model = "ph", corstr = "independence",
  Var = T, nboot = 100, esmax = 100, eps = 1e-06
)
print.smgeecure(bmt.ph.gee.ind)

# output the results
write.csv(cbind(bmt.ph.gee.ex$incidence,rbind(NA,bmt.ph.gee.ex$latency)),
          "bmt.ph.gee.ex.csv")
write.csv(cbind(bmt.ph.gee.ind$incidence,rbind(NA,bmt.ph.gee.ind$latency)),
          "bmt.ph.gee.ind.csv")

# ---- Fit the data using marginal semi-parametric marginal AFTMC model

# fit marginal semi-parametric marginal AFTMC model (exchangeable correlation)
set.seed(1)
bmt.aft.gee.ex <- smgeecure(
  formula = Surv(T2, d3) ~ factor(g) + Z8, cureform = ~ factor(g) + Z8,
  id = bmt$Z9, data = bmt, model = "aft", corstr = "exchangeable",
  Var = T,nboot = 100
)
print.smgeecure(bmt.aft.gee.ex)

# fit marginal semi-parametric AFTMC model (independence correlation)
set.seed(1)
bmt.aft.gee.ind <- smgeecure(
  formula = Surv(T2, d3) ~ factor(g) + Z8, cureform = ~ factor(g) + Z8,
  id = bmt$Z9, data = bmt, model = "aft", corstr = "independence",
  Var = T,nboot = 100
)
print.smgeecure(bmt.aft.gee.ind)

# output the results
write.csv(cbind(bmt.aft.gee.ex$incidence,rbind(NA,bmt.aft.gee.ex$latency)),
          "bmt.aft.gee.ex.csv")
write.csv(cbind(bmt.aft.gee.ind$incidence,rbind(NA,bmt.aft.gee.ind$latency)),
          "bmt.aft.gee.ind.csv")
}

}
\references{
Cai, C., Zou, Y., Peng, Y., and Zhang, J. (2012). smcure: An r-package for estimating semiparametric
mixture cure models. Computer methods and programs in biomedicine, 108(3):1255.

Zhang, J. and Peng, Y. (2007). A new estimation method for the semiparametric accelerated
failure time mixture cure model. Statistics in medicine, 26(16):3157–3171.

Niu, Y., Song, L., Liu, Y, and Peng, Y. (2018) Modeling clustered long-term survivors using marginal
mixture cure model. Biometrical Journal, doi: 10.1002/bjmj.201700114.
}
